Feature: Simple max width / height Feature: Auto max width / height based on scaleForBrowserSupport Feature: Simple test to help detect if browser is chrome.
I know that there is a maxImageSize for the pdfjs options.
I just do not know how that is implemented in their library.
Having the canvas be set to a max width or height is useful if you want
to reduce the load for certain browser enviroments where canvas size may
be restricted (safari has restrictions).
Reading through the api.d.ts file I found the isChrome flag, that will
be useful. Determines if we can use bmp ImageDecoder.